Update returned search_space of maturity and spend_limit filter
The two added lines in the maturity and the spend_limit filter are identical. With them, the filters, instead of removing the columns that correspond to replacements that should not be considered, turn the boolean values to False. 

This solves the related bug as it removes the possibility of an error if the filters are put in a row and makes the with_asset_technology filter work normally. 

Finally, this makes the maturity and spend_limit filters also more in line with other filters like same_fuel or similar_technology in how they manipulate the search space as they also just change the boolean values of the search_space DataArray, while keeping the dimensions identical.