Remove files autogenerated by Sphinx
We don't need simulation results etc. stored in the repo. They are used for the documentation, but are generated as needed anyway. They don't appear to be used for anything else and will likely just get out of sync when the code changes.

I've left the Results folders for the tutorial code for now as they are seemingly used by various notebooks. Maybe they can be removed in future.