Update test_docs_notebooks.py
Revoked changes to exclude the test in case input file is not there as per discussion for pull requests. 

Updated the manual to instruct users to run the default model on Muse to complete the installation. This ensures all tests pass (some may be skipped) and avoids including model results files in the codebase.